переписал под паттерн MVVM

+++ app/src/main/java/com/tests/lifehackstudiotest/domain/CompanyCard.kt
@@ -4,8 +4,8 @@ data class CompanyCard(
     val description: String,
     val id: String,
     val img: String,
-    val lat: Int,
-    val lon: Int,
+    val lat: Float,
+    val lon: Float,
     val name: String,
     val phone: String,
     val www: String
